requirements.txt
backports.statistics==0.1.0
google_api_python_client==2.80.0
google_auth_oauthlib==0.4.6
protobuf==4.22.0
requests==2.27.1
